Finish handle comment logic and response logic
